fix(activity): keep presence reporting out of render
React Doctor flagged two render-time ref writes in useClientPresence. A render
can be replayed or discarded, so a write from one that never commits would leak
into the next report; syncing both refs in an effect declared ahead of the
heartbeat keeps the ordering that made the render-time write correct.

The settings keys and the readings taken off them move to
utils/activitySettings.ts. App reads the staleness window without rendering the
pane, and a component file that also exports plain functions costs Fast Refresh
the ability to preserve component state.
