+++ app/src/components/SessionActivitySettings.test.tsx
@@ -1,6 +1,7 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import { describe, expect, it, vi } from 'vitest';
-import { SessionActivitySettings, activityStaleMs } from './SessionActivitySettings';
+import { SessionActivitySettings } from './SessionActivitySettings';
+import { activityStaleMs } from '../utils/activitySettings';
 
 const renderPane = (settings: Record<string, string> = {}) => {
   const onSetSetting = vi.fn();
